+++ model/schema_order_test.go
@@ -5,6 +5,7 @@ import "testing"
 func TestGetOrder(t *testing.T) {
 	// TODO: Schema.GetAscRows,GetAscRowsAt
 	// TODO: Schema.GetDescRows, GetDescRowsAt
+	// TODO: Schema.GetOrderedRows, GetOrderedRowsAt
 }
 
 func TestTopOrder(t *testing.T) {