+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+## v0.3.31 - Multi-Column Order
+  * Commit:
+  * Schema GetOrderedRows
 ## v0.3.30 - Secrets Package 
   * Commit: 2026-05-16 23:38
   * Secrets package